refactor: move all interfaces in 1 file
