+++ src/ts/types/interface.ts
@@ -11,6 +11,28 @@ export interface Router {
 }
 
 export interface RouterMatch {
-  route: Router;
-  result: RegExpMatchArray | null;
+    route: Router;
+    result: RegExpMatchArray | null;
+}
+
+export interface IProductCard {
+    id: number;
+    title: string;
+    description: string;
+    price: number;
+    discountPercentage: number;
+    rating: number;
+    stock: number;
+    brand: string;
+    category: string;
+    thumbnail: string;
+    images: string[];
+}
+
+export interface IProducts {
+    [key: string]: IProductCard;
+}
+
+export interface LoaderOption {
+    [key: string]: IProductCard[];
 }
